Introduce CampaignCellNodeDelegate
