+++ Source/CampaignDataSource.swift
@@ -13,19 +13,7 @@ extension CampaignDataSource: ASCollectionViewDataSource {
   public override func collectionView(collectionView: ASCollectionView!, nodeForItemAtIndexPath indexPath: NSIndexPath!) -> ASCellNode! {
     let cellNode: ASCellNode
 
-    if let delegate = delegate as? WallController,
-      post = delegate.post {
-        if indexPath.row > 0 {
-          cellNode = CommentCellNode(post: data[indexPath.row],
-            width: collectionView.frame.width, delegate)
-        } else {
-          cellNode = PostCellNode(post: data[indexPath.row],
-            width: collectionView.frame.width, delegate)
-        }
-    } else {
-      cellNode = PostCellNode(post: data[indexPath.row],
-        width: collectionView.frame.width, delegate)
-    }
+    cellNode = ASCellNode()
 
     return cellNode
   }
