+++ Source/Nodes/CampaignCellNodeDelegate.swift
@@ -1,0 +1,5 @@
+import Wall
+
+public protocol CampaignCellNodeDelegate: PostCellNodeDelegate {
+  var campaignConfig: Config { get }
+}
